backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "aiomysql>=0.2.0",
    "mysql-connector-python>=8.0.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pwdlib[argon2]>=0.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.21",
    "alembic>=1.18.0",
    "mysql-connector-python>=8.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
    "aiosqlite>=0.19.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["."]
omit = [
    "venv/*",
    "tests/*",
    "__pycache__/*",
    "*__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
